Refactor: Separating portfolio routes
