+++ src/pages/portfolios/[slug]/index.tsx
@@ -1,4 +1,4 @@
-export function Eijkman() {
+export function PortfolioDetails() {
   return (
     <div className="text-white flex justify-center items-center h-screen">
       Eijkman
